// swarm-demos/warehouse-factory-demo/src/generated/warehouse.ts
export interface Msg {
    /**
     * @generated from protobuf oneof: containedMessage
     */
    containedMessage: {
        oneofKind: "partReq";
        /**
         * @generated from protobuf field: warehouse_messages.PartReq partReq = 1;
         */
        partReq: PartReq;
    } | {
        oneofKind: "partOK";
        /**
         * @generated from protobuf field: warehouse_messages.PartOK partOK = 2;
         */
        partOK: PartOK;
    } | {
        oneofKind: "pos";
        /**
         * @generated from protobuf field: warehouse_messages.Pos pos = 3;
         */
        pos: Pos;
    } | {
        oneofKind: "closingTime";
        /**
         * @generated from protobuf field: warehouse_messages.ClosingTime closingTime = 4;
         */
        closingTime: ClosingTime;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message warehouse_messages.Meta
 */
export interface Meta {
    /**
     * @generated from protobuf field: bool isLocalEvent = 1;
     */
    isLocalEvent: boolean;
    /**
     * @generated from protobuf field: repeated string tags = 2;
     */
    tags: string[];
    /**
     * @generated from protobuf field: uint64 timestampMicros = 3;
     */
    timestampMicros: string;
    /**
     * @generated from protobuf field: uint32 lamport = 4;
     */
    lamport: number;
    /**
     * @generated from protobuf field: string appId = 5;
     */
    appId: string;
    /**
     * @generated from protobuf field: string eventId = 6;
     */
    eventId: string;
    /**
     * @generated from protobuf field: string stream = 7;
     */
    stream: string;
    /**
     * @generated from protobuf field: uint32 offset = 8;
     */
    offset: number;
}
/**
 * @generated from protobuf message warehouse_messages.PartReq
 */
export interface PartReq {
    /**
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * @generated from protobuf field: string lbj = 2;
     */
    lbj: string;
    /**
     * @generated from protobuf field: warehouse_messages.Meta meta = 3;
     */
    meta?: Meta;
}
/**
 * @generated from protobuf message warehouse_messages.PartOK
 */
export interface PartOK {
    /**
     * @generated from protobuf field: string part = 1;
     */
    part: string;
    /**
     * @generated from protobuf field: string lbj = 2;
     */
    lbj: string;
    /**
     * @generated from protobuf field: warehouse_messages.Meta meta = 3;
     */
    meta?: Meta;
}
/**
 * @generated from protobuf message warehouse_messages.Pos
 */
export interface Pos {
    /**
     * @generated from protobuf field: string position = 1;
     */
    position: string;
    /**
     * @generated from protobuf field: string part = 2;
     */
    part: string;
    /**
     * @generated from protobuf field: string lbj = 3;
     */
    lbj: string;
    /**
     * @generated from protobuf field: warehouse_messages.Meta meta = 4;
     */
    meta?: Meta;
}
/**
 * @generated from protobuf message warehouse_messages.ClosingTime
 */
export interface ClosingTime {
    /**
     * @generated from protobuf field: string timeOfDay = 1;
     */
    timeOfDay: string;
    /**
     * @generated from protobuf field: string lbj = 2;
     */
    lbj: string;
    /**
     * @generated from protobuf field: warehouse_messages.Meta meta = 3;
     */
    meta?: Meta;
}
// ...
export const Msg = new Msg$Type();
// ...
export const Meta = new Meta$Type();
// @generated message type with reflection information, may provide speed optimized methods
class PartReq$Type extends MessageType<PartReq> {
    constructor() {
        super("warehouse_messages.PartReq", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "lbj", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "meta", kind: "message", T: () => Meta }
        ]);
    }
    create(value?: PartialMessage<PartReq>): PartReq {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = "";
        message.lbj = "";
        if (value !== undefined)
            reflectionMergePartial<PartReq>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PartReq): PartReq {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string id */ 1:
                    message.id = reader.string();
                    break;
                case /* string lbj */ 2:
                    message.lbj = reader.string();
                    break;
                case /* warehouse_messages.Meta meta */ 3:
                    message.meta = Meta.internalBinaryRead(reader, reader.uint32(), options, message.meta);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PartReq, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string id = 1; */
        if (message.id !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.id);
        /* string lbj = 2; */
        if (message.lbj !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.lbj);
        /* warehouse_messages.Meta meta = 3; */
        if (message.meta)
            Meta.internalBinaryWrite(message.meta, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message warehouse_messages.PartReq
 */
export const PartReq = new PartReq$Type();
// @generated message type with reflection information, may provide speed optimized methods
class PartOK$Type extends MessageType<PartOK> {
    constructor() {
        super("warehouse_messages.PartOK", [
            { no: 1, name: "part", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "lbj", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "meta", kind: "message", T: () => Meta }
        ]);
    }
    create(value?: PartialMessage<PartOK>): PartOK {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.part = "";
        message.lbj = "";
        if (value !== undefined)
            reflectionMergePartial<PartOK>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PartOK): PartOK {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string part */ 1:
                    message.part = reader.string();
                    break;
                case /* string lbj */ 2:
                    message.lbj = reader.string();
                    break;
                case /* warehouse_messages.Meta meta */ 3:
                    message.meta = Meta.internalBinaryRead(reader, reader.uint32(), options, message.meta);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PartOK, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string part = 1; */
        if (message.part !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.part);
        /* string lbj = 2; */
        if (message.lbj !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.lbj);
        /* warehouse_messages.Meta meta = 3; */
        if (message.meta)
            Meta.internalBinaryWrite(message.meta, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message warehouse_messages.PartOK
 */
export const PartOK = new PartOK$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Pos$Type extends MessageType<Pos> {
    constructor() {
        super("warehouse_messages.Pos", [
            { no: 1, name: "position", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "part", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "lbj", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "meta", kind: "message", T: () => Meta }
        ]);
    }
    create(value?: PartialMessage<Pos>): Pos {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.position = "";
        message.part = "";
        message.lbj = "";
        if (value !== undefined)
            reflectionMergePartial<Pos>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Pos): Pos {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string position */ 1:
                    message.position = reader.string();
                    break;
                case /* string part */ 2:
                    message.part = reader.string();
                    break;
                case /* string lbj */ 3:
                    message.lbj = reader.string();
                    break;
                case /* warehouse_messages.Meta meta */ 4:
                    message.meta = Meta.internalBinaryRead(reader, reader.uint32(), options, message.meta);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Pos, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string position = 1; */
        if (message.position !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.position);
        /* string part = 2; */
        if (message.part !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.part);
        /* string lbj = 3; */
        if (message.lbj !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.lbj);
        /* warehouse_messages.Meta meta = 4; */
        if (message.meta)
            Meta.internalBinaryWrite(message.meta, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message warehouse_messages.Pos
 */
export const Pos = new Pos$Type();
// ...
export const ClosingTime = new ClosingTime$Type();